% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmodown.R
\docType{package}
\name{cosmodown}
\alias{cosmodown}
\title{cosmodown: Theses and dissertations for BYU}
\description{
This package creates theses/dissertations for Brigham Young University
}
\section{gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

